SAROJ-18 Fix to have exact match
With this current version there is a need of strict matching.

Social in sociala is not good enough as the ending a is not getting
matched afterwards and drags the whole offsets until the first "a"
which is blatantly wrong.

Now, the match is forced to be exact and else it will wait to find the
missing part. It will keep the duplication if there are issues with udpipe
tokens or spaces. This is taken care also in textReconstruction_process.